fix: share submit permission self (#590)
While creating a leave application, if the employee doesn't have a leave approver set, the document is shared with the leave applicant himself by the share API as the approver is `None`.

The Applicant ends up getting submit permission for leave application accidentally but cannot submit it as the "status" field is read-only and has perm level set.

Video:
https://github.com/frappe/hrms/assets/76736615/580b6ceb-9881-4801-b422-9e428ccf15f3