feat(Attendance Request):  add 'Include Holidays' checkbox (#1040)
In this PR, added a new feature that introduces the 'Include Holidays' checkbox in the attendance request form. This checkbox allows users to bypass the holiday validation when creating attendance records. It enhances the flexibility of the system by enabling manual attendance marking on holidays when required. Additionally, with this feature, employees can apply for compensatory leave if they work on a holiday.

https://github.com/frappe/hrms/assets/105106551/9256d331-2476-4a75-8440-ba8cc3c35b23




`no-docs`